prepare_ok: update fields (see full commit msg)
We need prepare_ok to work in a msg passing style
instead of in an RPC style. I think I now
understand why the paper does it this way.

Specifically, when a DoViewChange occurs and triggers a StartView, the
replicas and the new primary might end up with uncomitted ops in the
log, past the commit_number.

To allow the Primary to commit those
ops, the replicas will send a PrepareOk to the Primary (note that
this PrepareOk is not in reply to a Prepare msg in this case).
Because of this, using the RPC style (i.e. requiring in_reply_to) is
incompatible with this step of the protocol.

This step allows the new primary to finish processing requests made to
the earlier primary. I need to think more carefully about what this means
for the client table in our impl, but so far I see no issues.