bugfix: DoViewChange handler (see full commit msg)
There was a bug where I incorrectly obtained the latest op number.
The code was attempting to obtain it from a non-existent source, then
unwrapping the result. I was tired when I added that bug xD

Anyway it should now obtain the appropriate number.

This was discovered while running lin-kv-nemesis. Note that the test
still fails but due to an off-by-one error elsewhere. It's unclear if
this off-by-one error is due to our incomplete impl of GetState or due
to other causes.

Note that the missing part of our impl of GetState (as described in the
paper) will not work correctly under some configurations, as described
by Jack VanLightly. We will implement this part soon and test further.
It might be a rare edge case, or a common one. We'll see.