bugfix: StartView handler: op_number, commit_number
This commit fixes buggy handling of op_number and commit_number during
StartView change handling.

Prior to this fix, we incorrectly assumed that my_commit_number should
always be less than or equal to the commit_number in StartView.

We also forgot to trim the op_number when trimming the existing log.